Conflicts:
	libethereum/BlockInfo.cpp
	libethereum/Trie.h
	test/main.cpp

+++ libethereum/RLP.cpp
@@ -124,7 +124,7 @@ eth::uint RLP::items() const
 RLPStream& RLPStream::appendString(bytesConstRef _s)
 {
 	if (_s.size() < 0x38)
-		m_out.push_back(_s.size() | 0x40);
+		m_out.push_back((byte)(_s.size() | 0x40));
 	else
 		pushCount(_s.size(), 0x40);
 	uint os = m_out.size();
@@ -136,7 +136,7 @@ RLPStream& RLPStream::appendString(bytesConstRef _s)
 RLPStream& RLPStream::appendString(string const& _s)
 {
 	if (_s.size() < 0x38)
-		m_out.push_back(_s.size() | 0x40);
+		m_out.push_back((byte)(_s.size() | 0x40));
 	else
 		pushCount(_s.size(), 0x40);
 	uint os = m_out.size();
@@ -156,7 +156,7 @@ RLPStream& RLPStream::appendRaw(bytesConstRef _s)
 RLPStream& RLPStream::appendList(uint _count)
 {
 	if (_count < 0x38)
-		m_out.push_back(_count | 0x80);
+		m_out.push_back((byte)(_count | 0x80));
 	else
 		pushCount(_count, 0x80);
 	return *this;
@@ -165,11 +165,11 @@ RLPStream& RLPStream::appendList(uint _count)
 RLPStream& RLPStream::append(uint _i)
 {
 	if (_i < 0x18)
-		m_out.push_back(_i);
+		m_out.push_back((byte)_i);
 	else
 	{
 		auto br = bytesRequired(_i);
-		m_out.push_back(br + 0x17);	// max 8 bytes.
+		m_out.push_back((byte)(br + 0x17));	// max 8 bytes.
 		pushInt(_i, br);
 	}
 	return *this;
@@ -182,7 +182,7 @@ RLPStream& RLPStream::append(u160 _i)
 	else
 	{
 		auto br = bytesRequired(_i);
-		m_out.push_back(br + 0x17);	// max 8 bytes.
+		m_out.push_back((byte)(br + 0x17));	// max 8 bytes.
 		pushInt(_i, br);
 	}
 	return *this;
@@ -195,7 +195,7 @@ RLPStream& RLPStream::append(u256 _i)
 	else
 	{
 		auto br = bytesRequired(_i);
-		m_out.push_back(br + 0x17);	// max 8 bytes.
+		m_out.push_back((byte)(br + 0x17));	// max 8 bytes.
 		pushInt(_i, br);
 	}
 	return *this;
@@ -209,11 +209,11 @@ RLPStream& RLPStream::append(bigint _i)
 	{
 		uint br = bytesRequired(_i);
 		if (br <= 32)
-			m_out.push_back(bytesRequired(_i) + 0x17);	// max 32 bytes.
+			m_out.push_back((byte)(bytesRequired(_i) + 0x17));	// max 32 bytes.
 		else
 		{
 			auto brbr = bytesRequired(br);
-			m_out.push_back(0x37 + brbr);
+			m_out.push_back((byte)(0x37 + brbr));
 			pushInt(br, brbr);
 		}
 		pushInt(_i, br);
@@ -224,7 +224,7 @@ RLPStream& RLPStream::append(bigint _i)
 void RLPStream::pushCount(uint _count, byte _base)
 {
 	auto br = bytesRequired(_count);
-	m_out.push_back(br + 0x37 + _base);	// max 8 bytes.
+	m_out.push_back((byte)(br + 0x37 + _base));	// max 8 bytes.
 	pushInt(_count, br);
 }
